add hw13 doc and refactor errors
