+++ README.md
@@ -21,6 +21,7 @@
 8. [ Макросы](./docs/hw8.md)
 9. [ Матрицы](./docs/hw9.md)
 9. [ Транслируем идеи в код](./docs/hw10.md)
+13. [ Сервер и клиент для банка](./docs/hw13.md)
 
 </details>
