+++ docs/hw13.md
@@ -1,0 +1,35 @@
+# Домашнее задание: Сервер и клиент для банка
+
+### Цель: В этом домашнем задании необходимо разработать приложение, которое предоставит сетевой интерфейс для функционала банка, а также разработать клиентскую библиотеку.
+
+### Описание/Пошаговая инструкция выполнения домашнего задания
+
+#### Постановка задачи
+
+* Разработайте сервер, который предоставит сетевой интерфейс для всего функционала библиотеки, описанной в предыдущем
+  домашнем задании "Банк".
+* Создайте клиентскую библиотеку, которая обернет сетевой обмен в методы и предоставит доступ к сетевому интерфейсу
+  сервера.
+
+### Требования
+
+* Убедитесь, что код реализует весь описанный функционал.
+* Предоставьте пример использования клиентской библиотеки.
+* Задокументируйте все публичные методы клиентской библиотеки.
+* Библиотека клиента должна предоставлять информацию обо всех возникающих ошибках, включая сетевые ошибки.
+* Убедитесь, что команды cargo clippy и cargo fmt --check не выдают предупреждений и ошибок.
+* Рекомендуется разместить все пакеты (библиотеку банка, сервер банка и клиент банка) в одном workspace.
+* Для сериализации передаваемых данных вы можете использовать библиотеку serde с выбранным вами форматом: serde_json,
+  ron, toml.
+
+#### Дополнительная задача (по желанию): использовать gRPC для взаимодействия между клиентом и сервером.
+
+### Критерии оценки
+
+* "Принято": задание выполнено полностью.
+* "Возвращено на доработку": задание не выполнено полностью.
+
+---
+### Компетенции
+*  Алгоритмы и структуры данных
+*  Разработка клиент-серверных приложений
